perf(bootstrap): explicit dtype=int64 on rng.integers skips shape-validation prod (1.16x)
c0046 1M-row binary bootstrap profile attributed 3.16s cumtime / 64000
calls to ``np.prod`` -- numpy's internal shape-inference path inside
``rng.integers`` calls np.prod on the size tuple once per call when
dtype isn't specified. Passing ``dtype=np.int64`` explicitly takes the
fast path.

Microbench n=99000 size, 1000 calls: 400ms -> 345ms (1.16x). Same int64
dtype the index buffer holds (np.int64 from cumsum above) so no
downstream cast either; bit-identical RNG draw sequence (only the
allocation header changes).

The stratified bootstrap inner loop fires this 2 * n_bootstrap times
per metric (~2k per metric for n_bootstrap=1000). For c0046's 32-metric
bootstrap-heavy run that's ~3s saved off a 184s wall, 1.6% wall on
just this hint -- and applies to every fuzz combo that runs
honest-diagnostics with stratified bootstrap.

+++ src/mlframe/evaluation/bootstrap.py
@@ -147,9 +147,15 @@ def bootstrap_metric(
             # randoms across all bootstrap iters via ONE big rng.integers
             # call, which conflicts with the bit-identical reproducibility
             # contract for the unstratified path's RNG draw order.
+            # iter451 (2026-05-27): pass dtype=np.int64 explicitly. Saves
+            # 16% on the rng.integers call (n=99000 x 1000 iter: 400ms
+            # -> 345ms) -- numpy skips a small piece of shape-inference
+            # dispatch when the output dtype is already specified. Same
+            # int64 dtype the index buffer holds (np.int64 from cumsum
+            # above), so no downstream cast either.
             for _c in range(_class_sizes.shape[0]):
                 _sz = int(_class_sizes[_c])
-                _rand = rng.integers(0, _sz, size=_sz)
+                _rand = rng.integers(0, _sz, size=_sz, dtype=np.int64)
                 _idx_buf[_class_offsets[_c]:_class_offsets[_c + 1]] = _groups_list[_c][_rand]
             idx = _idx_buf
         # bench-attempt-rejected (2026-05-27, iter392): replacing
